apps/api/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn>=0.27.0,<1.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pyyaml>=6.0,<7.0.0
slowapi>=0.1.9,<1.0.0
